Controller as the contact delegate
Need to have the methods didBegin and DidEnd
from the Contact Delegate in the Controller
SuperClass, and override them in the child class,
otherwise the methods for the Contact delegate
would never be called. That is for being a child
class of the delegate, and those methods are
exposed to @objc.