Action resume popover
When showing Menu action sheet on an iPad
with a popover there was no button "Resume"
as popovers with action type "Cancel" is not
being presented. If the users clicks outside of
the popover it gets dismissed.

Resume action cannot be presented as a
"Default" actions, as there would be no
listener for when user clicks outside.

So we end up with two resume actions
when on iPad, but just 1 visible.